requirements.txt
insightface
onnxruntime
opencv-python
numpy
Pillow
imagehash
cinemagoer
icrawler
ddgs
requests
